Chicago Public Schools (CPS) - School Profile Information, School Year (SY) 2018-2019

School profile information for all schools in the Chicago Public School district for the school year 2018-2019.

data("cps_sy1819")

Format

A data frame of 660 observations - one record per school - on the following 81 variables:

school_id

current 6-digit unique ID assigned to each school

legacy_unit_id

outdated 4-digit unique ID assigned to each school

finance_id

5-digit unique ID assigned to each school (note that Englewood STEM High School has a value of 0 due to the fact that it is a new school set to open in the fall of 2019)

short_name

nickname of each school

long_name

formal name of each school

primary_category

primary grade levels served (ES for elementary, MS for middle, and HS for high school). Note that a school can possess any of these three values and still serve students the other two values.

is_high_school

T/F if school serves any 9-12th graders

is_middle_school

T/F if school serves any 6-8th graders

is_elementary_school

T/F if school serves any Kindergarten-5th graders

is_pre_school

T/F if school serves Pre-Kindergarteners

summary

text description of the school's mission and values

administrator_title

either Principal or Director

administrator

first and last name of school's administrator

address

school's address

city

city the school resides in (all Chicago)

state

state the school resides in (all Illinois)

zip

5-digit postal code used by the United States Postal Services

phone

10-digit phone number

fax

10-digit fax number used for telephonic transmission of scanned printed material

cps_school_profile

URL to a CPS provided web-page that provides an overview of the school

website

URL to school's home page

facebook

URL to school's Facebook page

twitter

URL to school's Twitter page

youtube

URL to a school's YouTube page

pinterest

URL to school's Pinterest page

attendance_boundaries

T/F if school has attendance boundaries (type in address here to see an example attendance boundary.)

grades_offered_all

grades offered by school, separated by a comma (i.e. 1,2,3)

grades_offered

grades offered by school, shortened by a dash (i.e. 1-3)

student_count_total

total number of students

student_count_low_income

number of low income students

student_count_special_ed

total number of students

student_count_english_learners

number of low income students

student_count_black

total number of African-American students

student_count_hispanic

total number of Hispanic students

student_count_white

total number of White students

student_count_asian

total number of Asian students

student_count_native_american

total number of Native-American students

student_count_other_ethnicity

total number of students whose race/ethnicity is 'Other'

student_count_asian_pacific_islander

total number of Asian/Pacific-Islander students

student_count_multi

total number of multi-racial/ethnic students

student_count_hawaiian_pacific_islander

total number of Hawaiian/Pacific-Islander students

student_count_ethnicity_not_available

total number of students whose ethnicity was not available

statistics_description

text description about the total student count, percent low income, percent diverse learners, and percent limited english learners

demographic_description

text description stating both the largest and second largest racial/ethnic demographic group

dress_code

T/F if school institutes a dress code

prek_school_day

pre-k school day description

kindergarten_school_day

kindergarten school days are all full days

school_hours

start and end time of school day (not a datetime)

freshman_start_end_time

start and end time of for freshman (not a datetime)

after_school_hours

start and end time of after school hours (not a datetime)

earliest_drop_off_time

earliest drop off time (not a datetime)

classroom_languages

non-English languages spoken in classroom, separated by a comma

bilingual_services

T/F if school offers bilingual services

refugee_services

T/F if school offers refugee services

title_1_eligible

T/F if school is eligible for Title 1 federal funds

preschool_inclusive

T/F if pre-school is inclusive

preschool_instructional

T/F if pre-school is instructional

significantly_modified

T/F if school offers services that provide students who need focused and very specific medical issues to supports for different learning styles and difficulties

hard_of_hearing

T/F if school offers curriculum for deaf and hard-of-hearing (HOH) students

visual_impairments

T/F if school offers curriculum for visually impaired students

transportation_bus

CTA bus routes near the school (only route numbers)

transportation_el

CTA 'L' train lines near the school (only the color(s) and not the exact stations)

transportation_metra

Metra train routes near the school

school_latitude

east-wide direction of the school (ranges from 41.65 to 42.02)

school_longitude

north-south direction of the school (ranges from -87.84 to -87.52)

college_enrollment_rate_school

college enrollment rates reflect the percent of CPS graduates who enrolled in college by the fall of their graduating year. CPS receives college enrollment data from the National Student Clearinghouse (NSC)

college_enrollment_rate_mean

the mean college enrollment rate is 68.2 percent

graduation_rate_school

high school graduation rate

graduation_rate_mean

the mean high school graduation rate is 78.2 percent

overall_rating

school quality rating policy (SQRP) rating, with Level 1+ being the highest performance and Level 3 being the lowest

rating_status

a school’s status determines who has decision-making power at the school level

rating_statement

text description of the value in overall_rating

classification_type

12 different types of school admission processes and curriculum

classification_description

text description of classification_type

school_year

School Year 2018-2019

network

District-run schools in CPS are organized into geographic networks, which provide administrative support, strategic direction, and leadership development to the schools within each network

is_gocps_participant

GoCPS allows families to learn, research, explore, and apply to nearly every CPS school and program through one online platform

is_gocps_prek

T/F if school serves Pre-Kindergarteners and is on GoCPS

is_gocps_elementary

T/F if school serves elementary grade students and is on GoCPS

is_gocps_high_school

T/F if school serves high school students and is on GoCPS

open_for_enrollment_date

open enrollment date

Details

This data frame does not contain all columns from the source file. The following columns were dropped due to irrelevance or all schools containing NA values:

average_act_school, mean_act, x_contact_name, x_contact_title, and closed_for_enrollment_date.

Source

This data frame comes from the City of Chicago Data Portal, Chicago Public Schools - School Profile Information SY1819. The file was last updated on December 29, 2018.

Examples

# load the necessary data -----
data("cps_sy1819")